+++ NEWS.md
@@ -1,3 +1,9 @@
+rphylopic (development version)
+==============
+
+* added functions for transforming PhyloPic silhouettes (flipping and rotating)
+* save_phylopic bg argument updated to be "transparent" by default
+
 rphylopic 1.0.0
 ==============